+++ src/dolphin/workflows/s1_disp.py
@@ -15,21 +15,21 @@ from dolphin.utils import get_max_memory_usage, set_num_threads
 
 from . import stitch_and_unwrap, wrapped_phase
 from ._utils import _create_burst_cfg, _remove_dir_if_empty
-from .config import Workflow
+from .config import DisplacementWorkflow
 
 
 @log_runtime
 def run(
-    cfg: Workflow,
+    cfg: DisplacementWorkflow,
     debug: bool = False,
 ):
     """Run the displacement workflow on a stack of SLCs.
 
     Parameters
     ----------
-    cfg : Workflow
-        [`Workflow`][dolphin.workflows.config.Workflow] object for controlling the
-        workflow.
+    cfg : DisplacementWorkflow
+        [`DisplacementWorkflow`][dolphin.workflows.config.DisplacementWorkflow] object
+        for controlling the workflow.
     debug : bool, optional
         Enable debug logging, by default False.
     """